fix tests

+++ internal/controller/dragonfly_instance.go
@@ -126,7 +126,10 @@ func (dfi *DragonflyInstance) configureReplica(ctx context.Context, pod *corev1.
 // checkReplicaRole returns true if the given pod is a replica and is connected to the correct master.
 func (dfi *DragonflyInstance) checkReplicaRole(ctx context.Context, pod *corev1.Pod, masterIp string) (bool, error) {
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
@@ -171,7 +174,10 @@ func (dfi *DragonflyInstance) checkReplicaRole(ctx context.Context, pod *corev1.
 // isReplicaStable returns true if the given replica is stable.
 func (dfi *DragonflyInstance) isReplicaStable(ctx context.Context, pod *corev1.Pod) (bool, error) {
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
@@ -443,7 +449,10 @@ func (dfi *DragonflyInstance) detectOldMasters(ctx context.Context, updateRevisi
 // replicaOf configures the pod as a replica to the given master instance
 func (dfi *DragonflyInstance) replicaOf(ctx context.Context, pod *corev1.Pod, masterIp string) error {
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
@@ -471,6 +480,7 @@ func (dfi *DragonflyInstance) replicaOf(ctx context.Context, pod *corev1.Pod, ma
 	}
 
 	dfi.log.Info("Marking pod role as replica", "pod", pod.Name, "masterIp", masterIp)
+	patch := client.MergeFrom(pod.DeepCopy())
 	pod.Labels[resources.RoleLabelKey] = resources.Replica
 	if pod.Annotations == nil {
 		pod.Annotations = make(map[string]string)
@@ -483,7 +493,7 @@ func (dfi *DragonflyInstance) replicaOf(ctx context.Context, pod *corev1.Pod, ma
 		pod.Labels[resources.MasterIpLabelKey] = masterIp
 	}
 
-	if err := dfi.client.Update(ctx, pod); err != nil {
+	if err := dfi.client.Patch(ctx, pod, patch); err != nil {
 		return fmt.Errorf("could not update replica metadata: %w", err)
 	}
 
@@ -493,7 +503,10 @@ func (dfi *DragonflyInstance) replicaOf(ctx context.Context, pod *corev1.Pod, ma
 // replicaOfNoOne configures the pod as a master along while updating other pods to be replicas
 func (dfi *DragonflyInstance) replicaOfNoOne(ctx context.Context, pod *corev1.Pod) error {
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
@@ -521,6 +534,7 @@ func (dfi *DragonflyInstance) replicaOfNoOne(ctx context.Context, pod *corev1.Po
 	masterIp := pod.Status.PodIP
 
 	dfi.log.Info("Marking pod role as master", "pod", pod.Name, "masterIp", masterIp)
+	patch := client.MergeFrom(pod.DeepCopy())
 	pod.Labels[resources.RoleLabelKey] = resources.Master
 	delete(pod.Labels, resources.MasterIpLabelKey)
 
@@ -529,7 +543,7 @@ func (dfi *DragonflyInstance) replicaOfNoOne(ctx context.Context, pod *corev1.Po
 	}
 	pod.Annotations[resources.MasterIpAnnotationKey] = masterIp
 
-	if err := dfi.client.Update(ctx, pod); err != nil {
+	if err := dfi.client.Patch(ctx, pod, patch); err != nil {
 		return err
 	}
 
@@ -598,11 +612,37 @@ func (dfi *DragonflyInstance) reconcileResources(ctx context.Context) error {
 			continue
 		}
 		// Update if specs differ
-		desired.SetResourceVersion(existing.GetResourceVersion())
-		if err = dfi.client.Update(ctx, desired); err != nil {
-			return fmt.Errorf("failed to update resource: %w", err)
+		patch := client.MergeFrom(existing.DeepCopyObject().(client.Object))
+		existingAnnotations := existing.GetAnnotations()
+		if existingAnnotations == nil {
+			existingAnnotations = make(map[string]string)
 		}
-		dfi.log.Info("updated resource", "resource", desired.GetName())
+		for k, v := range desired.GetAnnotations() {
+			existingAnnotations[k] = v
+		}
+		existing.SetAnnotations(existingAnnotations)
+
+		existingLabels := existing.GetLabels()
+		if existingLabels == nil {
+			existingLabels = make(map[string]string)
+		}
+		for k, v := range desired.GetLabels() {
+			existingLabels[k] = v
+		}
+		existing.SetLabels(existingLabels)
+
+		desiredV := reflect.ValueOf(desired).Elem()
+		existingV := reflect.ValueOf(existing).Elem()
+		desiredSpec := desiredV.FieldByName("Spec")
+		existingSpec := existingV.FieldByName("Spec")
+		if desiredSpec.IsValid() && existingSpec.IsValid() {
+			existingSpec.Set(desiredSpec)
+		}
+
+		if err = dfi.client.Patch(ctx, existing, patch); err != nil {
+			return fmt.Errorf("failed to patch resource: %w", err)
+		}
+		dfi.log.Info("patched resource", "resource", desired.GetName())
 	}
 	if dfi.df.Spec.Replicas < 2 {
 		if err = dfi.client.Delete(ctx, &policyv1.PodDisruptionBudget{
@@ -769,6 +809,11 @@ func (dfi *DragonflyInstance) allPodsHealthy(ctx context.Context, updateRevision
 	}
 
 	for _, pod := range pods.Items {
+		// Ignore pods that are already terminating
+		if isTerminating(&pod) {
+			continue
+		}
+
 		if !isPodOnLatestVersion(&pod, updateRevision) && !isTerminating(&pod) && !isRunningAndReady(&pod) {
 			dfi.log.Info("deleting failed to start pod", "pod", pod.Name)
 			if err := dfi.client.Delete(ctx, &pod); err != nil {
@@ -876,7 +921,10 @@ func (dfi *DragonflyInstance) replTakeover(ctx context.Context, newMaster *corev
 	dfi.log.Info("running REPLTAKEOVER on replica", "pod", newMaster.Name)
 
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(newMaster.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(newMaster.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
@@ -921,7 +969,10 @@ func (dfi *DragonflyInstance) getRedisRole(ctx context.Context, pod *corev1.Pod)
 		return "", fmt.Errorf("pod IP not available for %s", pod.Name)
 	}
 	redisClient := redis.NewClient(&redis.Options{
-		Addr: net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		Addr:         net.JoinHostPort(pod.Status.PodIP, strconv.Itoa(resources.DragonflyAdminPort)),
+		DialTimeout:  10 * time.Second,
+		ReadTimeout:  10 * time.Second,
+		WriteTimeout: 10 * time.Second,
 		MaintNotificationsConfig: &maintnotifications.Config{
 			Mode: maintnotifications.ModeDisabled,
 		},
